requirements.txt
gradio~=4.25.0
openai~=1.16.1
python-dotenv~=1.0.1
polarion~=1.3.0
certifi~=2024.2.2
requests~=2.31.0
termcolor~=2.4.0
langchain~=0.1.16
python-certifi-win32~=1.6.1